feat: student allocation suspsense + some small changes in animation pulse
