+++ apps/web/src/app/(dashboard)/(account)/account/_components/persons.tsx
@@ -75,7 +75,7 @@ function PersonsListFallback() {
             initials={"..."}
             className="bg-zinc-900 size-8 text-white"
           />
-          <div className="bg-zinc-300 rounded w-32 h-6 animate-pulse" />
+          <div className="bg-gray-200 rounded w-32 h-6 animate-pulse" />
         </GridListHeader>
         <DescriptionList className="px-6">
           <DescriptionTerm>NWD-id</DescriptionTerm>
