+++ apps/web/src/app/(dashboard)/(account)/profiel/[handle]/voortgang/[allocation-id]/_components/course-card.tsx
@@ -1,21 +1,13 @@
-import { PlusIcon } from "@heroicons/react/16/solid";
 import { notFound } from "next/navigation";
 import React, { Suspense } from "react";
-import { Button } from "~/app/(dashboard)/_components/button";
-import {
-  Checkbox,
-  CheckboxField,
-  CheckboxGroup,
-} from "~/app/(dashboard)/_components/checkbox";
+import { ModuleFallback } from "~/app/(dashboard)/(management)/locatie/[location]/cohorten/[cohort]/[student-allocation]/_components/student-module";
 import {
   DescriptionDetails,
   DescriptionList,
   DescriptionTerm,
 } from "~/app/(dashboard)/_components/description-list";
 import { Divider } from "~/app/(dashboard)/_components/divider";
-import { Label } from "~/app/(dashboard)/_components/fieldset";
-import { Strong, TextLink } from "~/app/(dashboard)/_components/text";
-import { Weight } from "~/app/_components/weight";
+import { TextLink } from "~/app/(dashboard)/_components/text";
 import {
   listCompetencyProgressInCohortForStudent,
   listCompletedCompetenciesByStudentCurriculumId,
@@ -119,49 +111,19 @@ function CourseCardFallback() {
       <DescriptionList>
         <DescriptionTerm>Programma</DescriptionTerm>
         <DescriptionDetails>
-          <span className="inline-block bg-zinc-200 rounded w-48 h-4.25 align-middle animate-pulse" />
+          <span className="inline-block bg-gray-200 rounded w-48 h-4.25 align-middle animate-pulse" />
         </DescriptionDetails>
 
         <DescriptionTerm>Vaartuig</DescriptionTerm>
         <DescriptionDetails>
-          <span className="inline-block bg-zinc-200 rounded w-32 h-4.25 align-middle animate-pulse [animation-delay:0.3s]" />
+          <span className="inline-block bg-gray-200 rounded w-32 h-4.25 align-middle animate-pulse [animation-delay:0.3s]" />
         </DescriptionDetails>
       </DescriptionList>
 
       <div className="mt-6">
         {[1, 2, 3].map((i, index) => (
           <React.Fragment key={`module-fallback-${i}`}>
-            {/* biome-ignore lint/a11y/useSemanticElements: <explanation> */}
-            <CheckboxGroup role="group">
-              <div className="flex justify-between items-center">
-                <div className="flex items-center gap-x-2">
-                  <Button plain>
-                    <PlusIcon />
-                  </Button>
-
-                  <CheckboxField disabled={true}>
-                    <Checkbox />
-                    <div className="flex gap-x-2">
-                      <Weight weight={1} />
-                      <Label>
-                        <Strong>
-                          <span
-                            className="inline-block bg-gray-300 rounded w-24 h-4.25 align-middle animate-pulse"
-                            style={{ animationDelay: `${index * 0.3}s` }}
-                          />
-                        </Strong>
-                      </Label>
-                    </div>
-                  </CheckboxField>
-                </div>
-                <div className="flex justify-end items-center gap-x-2">
-                  <span
-                    className="inline-flex bg-gray-200 rounded-md w-9.5 h-6 animate-pulse"
-                    style={{ animationDelay: `${index * 0.3 + 0.1}s` }}
-                  />
-                </div>
-              </div>
-            </CheckboxGroup>
+            <ModuleFallback index={index} />
 
             {index < 2 ? <Divider soft className="my-2.5" /> : null}
           </React.Fragment>
